+++ aldryn_django/management/commands/aldryn_update_s3_media_headers.py
@@ -1,0 +1,13 @@
+from django.core.files.storage import default_storage
+from django.core.management.base import BaseCommand, CommandError
+
+
+class Command(BaseCommand):
+    help = 'Ensure all headers for S3 uploaded files are up to date'
+
+    def handle(self, *args, **options):
+        if not hasattr(default_storage, 'update_headers'):
+            raise CommandError('The default media files storage does not '
+                               'support updating headers')
+        all_files, updated = default_storage.update_headers()
+        self.stdout.write('{}/{} files updated'.format(updated, all_files))